backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
tensorflow==2.15.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.1
seaborn==0.13.0
python-dotenv==1.0.0
httpx==0.25.1
websockets==12.0
paho-mqtt==1.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
